Mark and relocate built-in monitor tasks
